multi: add test client (#20)
* bond: implement flag.value interface.

This implements the flag.value interface in order to parse bond params from command line arguments.

* client: expose bond methods.

This adds an add bond method to the client for adding bonds dynamically and exposes the post bond method publicly.

* multi: add test client.

This adds the tatanka test client, the test client has a web interface for the client functionality and streams its logs.

* multi: implement flag.value via bond params wrapper.

This removes the bond params flag.value implementation in preference of scoping the implementation to where it's used: the executable. A
bond params wrapper is being used as result.