multi: Client connection maintenance
Adds a known nodes request to the tatanka node. This is periodically called
by the client to discover the addresses of the other tatanka nodes. If the
client's connection to the tatanka node is disconnected, it attempts to
connect to any other available tatanka node.

The push stream protocol is also updated. The client is now expected to
send an initial list of topics they want to be subscribed to.

The test UI is also updated to display the tatanka node that the client
is currently connected to.