tatanka: Refactor connection maintenance (#26)
Update manifest to a whitelist. Instead of separating bootstrap peers from
non-bootstrap peers, just allow the operator to specify addresses of the
other tatanka nodes which it knows about.

Updates the discovery protocol to just query all known addresses of another
peer.

Adds a whitelist handshake protocol which is performed after connection. If
the whitelists of the two tatanka nodes are not the same, the connection is
killed. All nodes in the mesh must have the same whitelist.

Add a `meshConnectionManager` type which attempts to maintain connections
to all other nodes in the mesh. ALl nodes are treated equally; there are
no special bootstrap nodes.